+++ backend/src/main/java/com/phungloccoffee/backend/dto/DoanhThuSanPhamResponse.java
@@ -1,8 +1,10 @@
 package com.phungloccoffee.backend.dto;
 
+import java.math.BigDecimal;
+
 public interface DoanhThuSanPhamResponse {
     String getMaSP();
     String getTenSP();
-    Long getTongSoLuongBan();
-    Double getTongDoanhThu();
+    Long getTongSoLuongBan(); 
+    BigDecimal getTongDoanhThu();
 }
